refactor: lift BamStatAccum into src/common/bam_stat_accum.rs
BamStatAccum gathers the read-level counters behind bam_stat and the
samtools writers. Its process_read takes only a record and a MAPQ cutoff,
so it is assay-agnostic and the dna pipeline will drive the same struct.
The merge_vec_arrays helper moves with it, being its only consumer.
rna::rseqc::accumulators re-exports the type.
